BingoCapsule, Grid 삭제. LoginPage Display Flex
